+++ .evergreen/scripts/compile.sh
@@ -146,6 +146,10 @@ cygwin)
 darwin*)
   cc_flags+=("${cc_flags_init[@]}")
   cxx_flags+=("${cxx_flags_init[@]}" -stdlib=libc++)
+
+  if [[ "${distro_id:?}" == macos-14* ]]; then
+    cxx_flags+=(-Wno-align-mismatch)
+  fi
   ;;
 linux*)
   cc_flags+=("${cc_flags_init[@]}")